requirements.txt

celery==5.3.0

crispy-bootstrap5==0.7

Django==4.2
django-crispy-forms==2.0
django-tinymce==3.6.1
django-utils-six==2.0

Pillow==9.5.0
redis==4.5.5

